refactor: ProductImage 컴포넌트 도메인 폴더로 이동
