+++ src/domains/components/ProductImage/ProductImage.styled.ts
